+++ pyalgos/sorting/__init__.py
@@ -1,3 +1,4 @@
 # -*- coding: utf-8 -*-
 
 from .insertion import insertion
+from .selection import selection