+++ CHANGELOG.rst
@@ -6,3 +6,4 @@ v0.1.0
 
 - First release.
 - Added ``insertion()`` method for implementing insertion sort algorithm.
+- Added ``selection()`` method for implementing selection sort algorithm.